pyproject.toml
[tool.poetry]
name = "monexo"
version = "0.1.0"
description = ""
authors = ["Mikhail Kozlov <kozlov_ma+nespamte@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
